Refactor code to seperate sync and async action of redux
