+++ src/containers/MealDetail.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { connect } from 'react-redux';
 import { useEffect } from 'react';
-import { fetchMealById } from '../redux/actions';
+import { fetchMealById } from '../redux/actions/thunk';
 import * as actionsType from '../redux/actions/actionTypes';
 import card from '../styles/card.css';
 import styles from '../styles/MealDetail.module.css';
